Working run
This is an upload of a working run of the software. Please take a look
at it and tell me if there are any mistakes I have made, or fix them
yourself :). We can delete the #define _XOPEN_SOURCE 600 before the
submission as I needed that to run on my environment.